% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilpalettes.R
\name{soil_palette}
\alias{soil_palette}
\title{Soil Palette Generator.}
\usage{
soil_palette(name, n, type = c("discrete", "continuous"), direction = 1)
}
\arguments{
\item{name}{Name of the color palette. Options are \code{durorthod}, \code{rendoll}, \code{paleustalf},
\code{natrudoll}, \code{alaquod2}, \code{alaquod}}

\item{n}{Number of colors in the palette. Palletes include 5 colors, which can be used discretely,
or if more are desired, used as a gradient. If omitted, n = length of palette.}

\item{type}{Usage of palette as "continuous" or "discrete". Continuous usage interpolates between colors to create
a scale of values. If omitted, function assumes continuous if n > length of palette, and discrete if n < length of palette.}

\item{direction}{(TODO)}
}
\value{
TODO
}
\description{
This function builds palettes based on soil profiles.
View photos for each palette \href{https://github.com/kaizadp/soilpalettes}{On Github}.
}
\examples{
soil_palette("durorthod",n=100,type="continuous")
soil_palette("durorthod",3)
soil_palette("durorthod", 50)
}
