% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilpalettes.R
\docType{data}
\name{soil_palettes}
\alias{soil_palettes}
\title{Complete list of palettes}
\format{An object of class \code{list} of length 6.}
\usage{
soil_palettes
}
\description{
Use \code{names(soil_palettes)} to view list of palette names.
Currently:  "durorthod" "rendoll"  "paleustalf"      "natrudoll"   "alaquod2"     "alaquod"
}
\keyword{datasets}
