% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilpalettes.R
\name{print.palette}
\alias{print.palette}
\title{Palette Print Function}
\usage{
\method{print}{palette}(x, ...)
}
\arguments{
\item{x}{something}

\item{...}{something}
}
\value{

}
\description{
Palette Print Function
}
